Move LibraryRegistryController to registration protocol blueprint and update references to new location
