+++ tests/test_scripts.py
@@ -13,7 +13,7 @@ from library_registry.model import (
 )
 from library_registry.model_helpers import (create, get_one)
 from library_registry.problem_details import INVALID_INTEGRATION_DOCUMENT
-from library_registry.registrar import LibraryRegistrar
+from library_registry.library_registration_protocol.registrar import LibraryRegistrar
 from library_registry.scripts import (
     AddLibraryScript,
     ConfigureEmailerScript,
